# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend untuk aplikasi Jagoan Bunda - Sistem monitoring tumbuh kembang anak",
        "title": "Jagoan Bunda API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jagoan Bunda Team",
            "email": "support@jagoanbunda.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/children/{childID}/anthropometry": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan semua data pengukuran anthropometry untuk anak tertentu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropometry"
                ],
                "summary": "Get semua record anthropometry anak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Child ID (UUID)",
                        "name": "childID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AnthropometryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data pengukuran anthropometry baru untuk anak (hanya Nakes)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropometry"
                ],
                "summary": "Tambah record anthropometry baru",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Child ID (UUID)",
                        "name": "childID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data anthropometry baru",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AnthropometryResponse"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.AnthropometryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/children/{childID}/anthropometry/{anthropometryID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendapatkan detail record anthropometry berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropometry"
                ],
                "summary": "Get record anthropometry by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Child ID (UUID)",
                        "name": "childID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Anthropometry ID",
                        "name": "anthropometryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AnthropometryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate data pengukuran anthropometry anak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropometry"
                ],
                "summary": "Update record anthropometry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Child ID (UUID)",
                        "name": "childID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Anthropometry ID",
                        "name": "anthropometryID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data anthropometry yang diupdate",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AnthropometryResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AnthropometryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus record anthropometry berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anthropometry"
                ],
                "summary": "Hapus record anthropometry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Child ID (UUID)",
                        "name": "childID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Anthropometry ID",
                        "name": "anthropometryID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AnthropometryResponse": {
            "type": "object",
            "properties": {
                "age_months": {
                    "type": "integer"
                },
                "head_circumference": {
                    "type": "number"
                },
                "height": {
                    "type": "number"
                },
                "status_bbtb": {
                    "type": "number"
                },
                "status_bbu": {
                    "type": "number"
                },
                "status_tbu": {
                    "type": "number"
                },
                "time": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                },
                "zscore_bbtb": {
                    "type": "number"
                },
                "zscore_bbu": {
                    "type": "number"
                },
                "zscore_tbu": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}